Revision 1.1
Added bit C3 - Start PB latch. When the start button is released, the relay will still be activated, because of the latch C3
Added X103 in Rung 1. - each of the subroutines needs a trigger to start. That is accomplished by the X103 momentary input foot pedal.
Added in Rung 4 NO Momentary Cycle Stop PB (X102) in both subroutines